Python 3 support w/ tests

+++ ple/games/catcher.py
@@ -1,6 +1,6 @@
 import sys
 import pygame
-from utils import percent_round_int
+from .utils import percent_round_int
 
 from ple.games import base
 from pygame.constants import K_a, K_d
@@ -101,7 +101,7 @@ class Fruit(pygame.sprite.Sprite):
         y = self.rng.choice(
             range(
                 self.size,
-                self.SCREEN_HEIGHT / 2,
+                int(self.SCREEN_HEIGHT / 2),
                 self.size))
 
         self.rect.center = (x, -1 * y)
